requirements.txt
ftfy==6.0.3
grad-cam==1.3.7
librosa==0.8.1
numpy==1.19.2
opencv-python==4.5.4.60
pandas
Pillow==8.4.0
pytorch-lightning==1.1.3
regex==2021.10.8
scikit-learn==1.0
scikit-video==1.1.11
scipy==1.7.1
torch==1.11.0
torchaudio==0.11.0
torchvision==0.12.0
tqdm==4.62.3
ttach==0.0.3
protobuf==3.20.3
